requirements.txt
telethon
requests
python-dotenv
requests
